refactor: 프로젝트 구조 개선 (Provider 모듈화)
